Cleanup packages, split files, etc
