+++ pkg/render/node_test.go
@@ -2284,196 +2284,6 @@ var _ = Describe("Node rendering tests", func() {
 				Expect(ns["projectcalico.org/operator-node-migration"]).To(Equal("migrated"))
 			})
 
-			trueValue := true
-			falseValue := false
-
-			// TODO: Re-write these tests as part of pool controller.
-			DescribeTable("test IP Pool configuration",
-				func(pool operatorv1.IPPool, expect map[string]string) {
-					cfg.IPPools = []operatorv1.IPPool{pool}
-					component := render.Node(&cfg)
-					Expect(component.ResolveImages(nil)).To(BeNil())
-					resources, _ := component.Objects()
-					Expect(len(resources)).To(Equal(defaultNumExpectedResources))
-
-					dsResource := rtest.GetResource(resources, "calico-node", "calico-system", "apps", "v1", "DaemonSet")
-					Expect(dsResource).ToNot(BeNil())
-
-					// The DaemonSet should have the correct configuration.
-					ds := dsResource.(*appsv1.DaemonSet)
-					rtest.ExpectEnv(ds.Spec.Template.Spec.Containers[0].Env, "NO_DEFAULT_POOLS", "true")
-				},
-
-				Entry("Default pool",
-					operatorv1.IPPool{
-						CIDR: "192.168.0.0/16",
-					},
-					map[string]string{
-						"CALICO_IPV4POOL_CIDR": "192.168.0.0/16",
-						"CALICO_IPV4POOL_IPIP": "Always",
-					}),
-
-				Entry("Pool with nat outgoing disabled",
-					operatorv1.IPPool{
-						CIDR:        "172.16.0.0/24",
-						NATOutgoing: "Disabled",
-					},
-					map[string]string{
-						"CALICO_IPV4POOL_CIDR":         "172.16.0.0/24",
-						"CALICO_IPV4POOL_IPIP":         "Always",
-						"CALICO_IPV4POOL_NAT_OUTGOING": "false",
-					}),
-				Entry("Pool with nat outgoing enabled",
-					operatorv1.IPPool{
-						CIDR:        "172.16.0.0/24",
-						NATOutgoing: "Enabled",
-					},
-					map[string]string{
-						"CALICO_IPV4POOL_CIDR": "172.16.0.0/24",
-						"CALICO_IPV4POOL_IPIP": "Always",
-						// Enabled is the default so we don't set
-						// NAT_OUTGOING if it is enabled.
-					}),
-				Entry("Pool with VXLAN enabled (IPv6)",
-					operatorv1.IPPool{
-						CIDR:          "fc00::/48",
-						Encapsulation: operatorv1.EncapsulationVXLAN,
-					},
-					map[string]string{
-						"CALICO_IPV6POOL_CIDR":  "fc00::/48",
-						"CALICO_IPV6POOL_VXLAN": "Always",
-					}),
-				Entry("Pool with VXLAN cross subnet enabled (IPv6)",
-					operatorv1.IPPool{
-						CIDR:          "fc00::/48",
-						Encapsulation: operatorv1.EncapsulationVXLANCrossSubnet,
-					},
-					map[string]string{
-						"CALICO_IPV6POOL_CIDR":  "fc00::/48",
-						"CALICO_IPV6POOL_VXLAN": "CrossSubnet",
-					}),
-				Entry("Pool with nat outgoing disabled (IPv6)",
-					operatorv1.IPPool{
-						CIDR:        "fc00::/48",
-						NATOutgoing: "Disabled",
-					},
-					map[string]string{
-						"CALICO_IPV6POOL_CIDR": "fc00::/48",
-						// Disabled is the default so we don't set
-						// NAT_OUTGOING if it is disabled.
-					}),
-				Entry("Pool with nat outgoing enabled (IPv6)",
-					operatorv1.IPPool{
-						CIDR:        "fc00::/48",
-						NATOutgoing: "Enabled",
-					},
-					map[string]string{
-						"CALICO_IPV6POOL_CIDR":         "fc00::/48",
-						"CALICO_IPV6POOL_NAT_OUTGOING": "true",
-					}),
-				Entry("Pool with CrossSubnet",
-					operatorv1.IPPool{
-						CIDR:          "172.16.0.0/24",
-						Encapsulation: operatorv1.EncapsulationIPIPCrossSubnet,
-					},
-					map[string]string{
-						"CALICO_IPV4POOL_CIDR": "172.16.0.0/24",
-						"CALICO_IPV4POOL_IPIP": "CrossSubnet",
-					}),
-				Entry("Pool with VXLAN",
-					operatorv1.IPPool{
-						CIDR:          "172.16.0.0/24",
-						Encapsulation: operatorv1.EncapsulationVXLAN,
-					},
-					map[string]string{
-						"CALICO_IPV4POOL_CIDR":  "172.16.0.0/24",
-						"CALICO_IPV4POOL_VXLAN": "Always",
-					}),
-				Entry("Pool with VXLANCrossSubnet",
-					operatorv1.IPPool{
-						CIDR:          "172.16.0.0/24",
-						Encapsulation: operatorv1.EncapsulationVXLANCrossSubnet,
-					},
-					map[string]string{
-						"CALICO_IPV4POOL_CIDR":  "172.16.0.0/24",
-						"CALICO_IPV4POOL_VXLAN": "CrossSubnet",
-					}),
-				Entry("Pool with no encapsulation",
-					operatorv1.IPPool{
-						CIDR:          "172.16.0.0/24",
-						Encapsulation: operatorv1.EncapsulationNone,
-					},
-					map[string]string{
-						"CALICO_IPV4POOL_CIDR": "172.16.0.0/24",
-						"CALICO_IPV4POOL_IPIP": "Never",
-					}),
-				Entry("Pool with node selector",
-					operatorv1.IPPool{
-						CIDR:         "172.16.0.0/24",
-						NodeSelector: "has(thiskey)",
-					},
-					map[string]string{
-						"CALICO_IPV4POOL_CIDR":          "172.16.0.0/24",
-						"CALICO_IPV4POOL_IPIP":          "Always",
-						"CALICO_IPV4POOL_NODE_SELECTOR": "has(thiskey)",
-					}),
-				Entry("Pool with v4 disable BGP export set to true",
-					operatorv1.IPPool{
-						CIDR:             "172.16.0.0/24",
-						DisableBGPExport: &trueValue,
-					},
-					map[string]string{
-						"CALICO_IPV4POOL_CIDR":               "172.16.0.0/24",
-						"CALICO_IPV4POOL_IPIP":               "Always",
-						"CALICO_IPV4POOL_DISABLE_BGP_EXPORT": "true",
-					}),
-				Entry("Pool with v4 disable BGP export set to false",
-					operatorv1.IPPool{
-						CIDR:             "172.16.0.0/24",
-						DisableBGPExport: &falseValue,
-					},
-					map[string]string{
-						"CALICO_IPV4POOL_CIDR":               "172.16.0.0/24",
-						"CALICO_IPV4POOL_IPIP":               "Always",
-						"CALICO_IPV4POOL_DISABLE_BGP_EXPORT": "false",
-					}),
-				Entry("Pool with v6 disable BGP export set to true",
-					operatorv1.IPPool{
-						CIDR:             "fc00::/48",
-						DisableBGPExport: &trueValue,
-					},
-					map[string]string{
-						"CALICO_IPV6POOL_CIDR":               "fc00::/48",
-						"CALICO_IPV6POOL_IPIP":               "Always",
-						"CALICO_IPV6POOL_DISABLE_BGP_EXPORT": "true",
-					}),
-				Entry("Pool with v6 disable BGP export set to false",
-					operatorv1.IPPool{
-						CIDR:             "fc00::/48",
-						DisableBGPExport: &falseValue,
-					},
-					map[string]string{
-						"CALICO_IPV6POOL_CIDR":               "fc00::/48",
-						"CALICO_IPV6POOL_IPIP":               "Always",
-						"CALICO_IPV6POOL_DISABLE_BGP_EXPORT": "false",
-					}),
-				Entry("Pool with all fields set",
-					operatorv1.IPPool{
-						CIDR:             "172.16.0.0/24",
-						Encapsulation:    operatorv1.EncapsulationIPIP,
-						NATOutgoing:      "Disabled",
-						NodeSelector:     "has(thiskey)",
-						DisableBGPExport: &trueValue,
-					},
-					map[string]string{
-						"CALICO_IPV4POOL_CIDR":               "172.16.0.0/24",
-						"CALICO_IPV4POOL_IPIP":               "Always",
-						"CALICO_IPV4POOL_NAT_OUTGOING":       "false",
-						"CALICO_IPV4POOL_NODE_SELECTOR":      "has(thiskey)",
-						"CALICO_IPV4POOL_DISABLE_BGP_EXPORT": "true",
-					}),
-			)
-
 			It("should not enable prometheus metrics if NodeMetricsPort is nil", func() {
 				defaultInstance.Variant = operatorv1.TigeraSecureEnterprise
 				defaultInstance.NodeMetricsPort = nil
@@ -2691,7 +2501,8 @@ var _ = Describe("Node rendering tests", func() {
 					{
 						Key:   "net.ipv4.tcp_keepalive_intvl",
 						Value: "15",
-					}, {
+					},
+					{
 						Key:   "net.ipv4.tcp_keepalive_probes",
 						Value: "6",
 					},
